Fix shurikens, throwing knives, and javelins fitting in the quiver bauble slot
